refactor: moved result types to session_types.py
